requirements.txt
pyyaml
pyperclip
cryptography